update layer inputs by layer node; remove redundant data structures
