+++ Stitch/Graph/Node/Layer/ViewModel/LayerNodeReader.swift
@@ -10,6 +10,13 @@ import Foundation
 protocol LayerNodeReader {
     @MainActor func getLayerInputObserver(_ layerInput: LayerInputPort) -> LayerInputObserver
     @MainActor var layerGroupId: NodeId? { get }
+    @MainActor var allLayerInputObservers: [LayerInputObserver] { get }
 }
 
-extension LayerNodeViewModel: LayerNodeReader { }
+extension LayerNodeViewModel: LayerNodeReader {
+    @MainActor var allLayerInputObservers: [LayerInputObserver] {
+        self.layer.layerGraphNode.inputDefinitions.reduce(into: .init()) { partialResult, port in
+            partialResult.append(self[keyPath: port.layerNodeKeyPath])
+        }
+    }
+}
